apps/api/pyproject.toml
[project]
name = "agent-control-plane-api"
version = "0.1.0"
description = "FastAPI control plane for agents, secrets, and lightweight tasks"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "pydantic-settings>=2.10.1",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
]

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
